Blender refactor (#122)
* moved prepare_blender_object to new blender_utils.py file

* moved various image IO functions to kubric.utils

* updated blender container to blender293

This simplifies the build quite a bit because the segfault patch is obsolete, and we can simply use the system python3.9-dev package instead of compiling it ourselves

* moved compute_bboxes to post_processing.py

* moved PathLike and AddAssetFunction to new custom_types file

* moved compositor node setup for EXR files out into blender_utils.py

* fixed pip and removed mathutils requirement (is installed as part of bpy)

* minor adjustments to new blender 2.93 API

* moved clear_and_reset and activate_render_passes to blender_utils.py

* moved blender-specific (exr) postprocessing to blender_utils.py

* small fix for type linter and added back compatibility with 2.91
