+++ docker/Blender.Dockerfile
@@ -1,10 +1,10 @@
 # Compiles a docker image for blender w/ "import bpy support"
 # 
 # Compilation happens in two stages:
-# 1) Compiles python and blender from source. 
-# 2) Installs previously built python and bpy module along with other dependencies in a fresh image.
-# This two stage process makes the final image much smaller because it doesn't include the files
-# and dependencies of the build process.
+# 1) Compiles blender from source.
+# 2) Installs previously built bpy module along with other dependencies in a fresh image.
+# This two stage process reduces the size of the final image because it doesn't include
+# the files and dependencies of the build process.
 
 # #################################################################################################
 # Stage 1
@@ -20,18 +20,11 @@ WORKDIR /blenderpy
 
 # --- Install package dependencies
 RUN apt-get update --yes --fix-missing && \
-    apt-get install --yes --quiet --no-install-recommends --reinstall \
-      # basic dependencies
+    apt-get install --yes --quiet --no-install-recommends \
+      python3.9-dev \
       build-essential \
-      wget \
-      curl \
       ca-certificates \
-      checkinstall \
-      # for GIF creation
-      imagemagick \
-      # OpenEXR
       libopenexr-dev \
-      # blender dependencies
       cmake \
       git \
       libffi-dev \
@@ -43,34 +36,16 @@ RUN apt-get update --yes --fix-missing && \
       libxrandr-dev \
       libxinerama-dev \
       libglew-dev \
-      subversion \
-      zlib1g-dev \
-      # further (optional) python build dependencies
-      libbz2-dev \
-      libgdbm-dev \
-      liblzma-dev \
-      libncursesw5-dev \
-      libreadline-dev \
-      libsqlite3-dev \
-      #tk-dev \  # installs libpng-dev which leads to blender linking errors
-      uuid-dev
+      subversion
 
-
-# --- Compile Python 3.7 from source (not available anymore in Ubuntu 20.04)
-
-RUN wget -nv https://www.python.org/ftp/python/3.7.9/Python-3.7.9.tar.xz -O Python.tar.xz && \
-    tar xJf Python.tar.xz && \
-    rm -f Python.tar.xy &&\
-    cd Python-3.7.9 && \
-    ./configure --enable-optimizations && \
-    make install && \
-    checkinstall --default
-
+# make python3.9 the default python
+RUN update-alternatives --install /usr/bin/python python /usr/bin/python3.9 10 && \
+    update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.9 10
 
 # --- Clone and compile Blender
 
 # RUN git clone https://git.blender.org/blender.git
-RUN git clone https://github.com/blender/blender.git --branch blender-v2.91-release --depth 1
+RUN git clone https://github.com/blender/blender.git --branch blender-v2.93-release --depth 1
 
 RUN mkdir lib && \
     cd lib && \
@@ -78,13 +53,6 @@ RUN mkdir lib && \
 
 RUN cd blender && \
     make update
-
-# Patch to fix segfault on exit problem
-# https://developer.blender.org/T82675
-# https://developer.blender.org/rB87d3f4aff3225104cbb8be41ac0339c6a1cd9a85
-# TODO: remove once we are using Blender 2.92
-COPY ./docker/segfault_bug_patch.txt /blenderpy/blender
-RUN cd blender && patch -p1 < segfault_bug_patch.txt
 
 # fix an annoying (no-consequence) bpy shutdown error
 # see https://github.com/google-research/kubric/issues/65
@@ -112,6 +80,8 @@ ENV LANG C.UTF-8
 # TODO: probably do not need all of them, or at least not in their dev version
 RUN apt-get update --yes --fix-missing && \
     apt-get install --yes --quiet --no-install-recommends --reinstall \
+      python3.9-dev \
+      python3.9-distutils \
       build-essential \
       # for GIF creation
       imagemagick \
@@ -140,9 +110,15 @@ RUN apt-get update --yes --fix-missing && \
       #tk-dev \  # installs libpng-dev which leads to blender linking errors
       uuid-dev
 
-COPY --from=build /blenderpy/Python-3.7.9/python_3.7.9-1_amd64.deb .
-RUN dpkg -i python_3.7.9-1_amd64.deb && \
-    rm -f python_3.7.9-1_amd64.deb
+# make python3.9 the default python and python3
+RUN update-alternatives --install /usr/bin/python python /usr/bin/python3.9 10 && \
+    update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.9 10
 
-COPY --from=build /blenderpy/build_linux_bpy/bin/bpy.so /usr/local/lib/python3.7/site-packages
-COPY --from=build /blenderpy/lib/linux_centos7_x86_64/python/lib/python3.7/site-packages/2.91 /usr/local/lib/python3.7/site-packages/2.91
+# install pip for python 3.9
+RUN curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py && \
+    python3.9 get-pip.py && \
+    rm get-pip.py
+
+
+COPY --from=build /blenderpy/build_linux_bpy/bin/bpy.so /usr/local/lib/python3.9/dist-packages/
+COPY --from=build /blenderpy/lib/linux_centos7_x86_64/python/lib/python3.9/site-packages/2.93 /usr/local/lib/python3.9/dist-packages/2.93
