+++ docker/Kubruntu.Dockerfile
@@ -13,15 +13,16 @@ WORKDIR /kubric
 
 # --- Install Python dependencies
 COPY requirements.txt .
-RUN python3 -m ensurepip
-RUN pip3 install --upgrade pip wheel
-RUN pip3 install --upgrade --force-reinstall -r requirements.txt
-RUN rm -f requirements.txt
+
+RUN pip install --upgrade pip wheel && \
+    pip install --upgrade --force-reinstall -r requirements.txt && \
+    rm -f requirements.txt
 
 # --- Silences tensorflow
 ENV TF_CPP_MIN_LOG_LEVEL="3"
 
 # --- Install Kubric
 COPY dist/kubric*.whl .
-RUN pip3 install `ls kubric*.whl`
-RUN rm -f kubric*.whl
+
+RUN pip3 install `ls kubric*.whl` && \
+    rm -f kubric*.whl
