--- docker/segfault_bug_patch.txt
@@ -1,54 +1,0 @@
-diff --git a/source/blender/python/intern/bpy_interface.c b/source/blender/python/intern/bpy_interface.c
-index d72c2cc..b3ae8d3 100644
---- a/source/blender/python/intern/bpy_interface.c
-+++ b/source/blender/python/intern/bpy_interface.c
-@@ -426,6 +426,13 @@ void BPY_python_start(bContext *C, int argc, const char **argv)
-   py_tstate = PyGILState_GetThisThreadState();
-   PyEval_ReleaseThread(py_tstate);
- #endif
-+
-+#ifdef WITH_PYTHON_MODULE
-+  /* Disable all add-ons at exit, not essential, it just avoids resource leaks, see T71362. */
-+  BPY_run_string_eval(C,
-+                      (const char *[]){"atexit", "addon_utils", NULL},
-+                      "atexit.register(addon_utils.disable_all)");
-+#endif
- }
-
- void BPY_python_end(void)
-diff --git a/source/blender/windowmanager/CMakeLists.txt b/source/blender/windowmanager/CMakeLists.txt
-index 07e962a..21cc293 100644
---- a/source/blender/windowmanager/CMakeLists.txt
-+++ b/source/blender/windowmanager/CMakeLists.txt
-@@ -165,6 +165,9 @@ if(WITH_PYTHON)
-     ../python
-   )
-   add_definitions(-DWITH_PYTHON)
-+  if(WITH_PYTHON_MODULE)
-+    add_definitions(-DWITH_PYTHON_MODULE)
-+  endif()
- endif()
-
- if(WITH_BUILDINFO)
-diff --git a/source/blender/windowmanager/intern/wm_init_exit.c b/source/blender/windowmanager/intern/wm_init_exit.c
-index ac9d384..9a8c7b2 100644
---- a/source/blender/windowmanager/intern/wm_init_exit.c
-+++ b/source/blender/windowmanager/intern/wm_init_exit.c
-@@ -521,11 +521,15 @@ void WM_exit_ex(bContext *C, const bool do_python)
-     }
-   }
-
--#ifdef WITH_PYTHON
-+#if defined(WITH_PYTHON) && !defined(WITH_PYTHON_MODULE)
-   /* Without this, we there isn't a good way to manage false-positive resource leaks
-    * where a #PyObject references memory allocated with guarded-alloc, T71362.
-    *
--   * This allows add-ons to free resources when unregistered (which is good practice anyway). */
-+   * This allows add-ons to free resources when unregistered (which is good practice anyway).
-+   *
-+   * Don't run this code when built as a Python module as this runs when Python is in the
-+   * process of shutting down, where running a snippet like this will crash, see T82675.
-+   * Instead use the `atexit` module, installed by #BPY_python_start */
-   BPY_run_string_eval(C, (const char *[]){"addon_utils", NULL}, "addon_utils.disable_all()");
- #endif
-
