+++ kubric/custom_types.py
@@ -1,0 +1,26 @@
+# Copyright 2021 The Kubric Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import Any, Callable, Union
+
+import tensorflow_datasets.public_api as tfds
+
+# noinspection PyUnresolvedReferences
+from kubric import core
+
+__all__ = ["AddAssetFunction", "PathLike"]
+
+AddAssetFunction = Callable[["core.View", "core.Asset"], Any]
+
+PathLike = Union[str, tfds.core.ReadWritePath]
