+++ requirements.txt
@@ -5,7 +5,6 @@ cloudml-hypertune
 dataclasses
 google-cloud-storage
 imageio
-mathutils
 munch
 numpy>=1.17
 pandas
